+++ wnp-server/src/test/java/com/wonnapark/wnpserver/episode/presentation/CommonEpisodeControllerTest.java
@@ -1,11 +1,11 @@
-package com.wonnapark.wnpserver.domain.episode.presentation;
+package com.wonnapark.wnpserver.episode.presentation;
 
 import com.wonnapark.wnpserver.auth.application.AuthenticationResolver;
 import com.wonnapark.wnpserver.episode.Episode;
+import com.wonnapark.wnpserver.episode.EpisodeFixtures;
 import com.wonnapark.wnpserver.episode.application.EpisodeFindUseCase;
 import com.wonnapark.wnpserver.episode.dto.response.EpisodeDetailFormResponse;
 import com.wonnapark.wnpserver.episode.dto.response.EpisodeListFormResponse;
-import com.wonnapark.wnpserver.episode.presentation.CommonEpisodeController;
 import com.wonnapark.wnpserver.global.auth.AuthorizedArgumentResolver;
 import com.wonnapark.wnpserver.global.auth.JwtAuthenticationInterceptor;
 import com.wonnapark.wnpserver.webtoon.Webtoon;
@@ -24,10 +24,6 @@ import java.util.List;
 
 import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.document;
 import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.resourceDetails;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createEpisode;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createEpisodes;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createPageable;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createWebtoon;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
@@ -63,8 +59,8 @@ class CommonEpisodeControllerTest {
     @DisplayName("에피소드 ID로 에피소드 상세 정보를 정상적으로 가져올 수 있다.")
     void findEpisodeDetailForm() throws Exception {
         // given
-        Webtoon webtoon = createWebtoon();
-        Episode episode = createEpisode(webtoon);
+        Webtoon webtoon = EpisodeFixtures.createWebtoon();
+        Episode episode = EpisodeFixtures.createEpisode(webtoon);
         given(episodeFindUseCase.findEpisodeDetailForm(episode.getId())).willReturn(EpisodeDetailFormResponse.from(episode));
         given(jwtAuthenticationInterceptor.preHandle(any(), any(), any())).willReturn(true);
         // when // then
@@ -90,9 +86,9 @@ class CommonEpisodeControllerTest {
     @DisplayName("웹툰 ID로 에피소드 리스트 정보를 정상적으로 가져올 수 있다.")
     void findEpisodeListForm() throws Exception {
         // given
-        Pageable pageable = createPageable();
-        Webtoon webtoon = createWebtoon();
-        List<Episode> episodes = createEpisodes(webtoon);
+        Pageable pageable = EpisodeFixtures.createPageable();
+        Webtoon webtoon = EpisodeFixtures.createWebtoon();
+        List<Episode> episodes = EpisodeFixtures.createEpisodes(webtoon);
         given(jwtAuthenticationInterceptor.preHandle(any(), any(), any())).willReturn(true);
         given(episodeFindUseCase.findEpisodeListForm(webtoon.getId(), pageable))
                 .willReturn(new PageImpl<>(episodes, pageable, episodes.size()).map(EpisodeListFormResponse::from));
