+++ wnp-server/src/test/java/com/wonnapark/wnpserver/episode/presentation/UserEpisodeControllerTest.java
@@ -1,11 +1,11 @@
-package com.wonnapark.wnpserver.domain.episode.presentation;
+package com.wonnapark.wnpserver.episode.presentation;
 
 import com.wonnapark.wnpserver.auth.application.AuthenticationResolver;
 import com.wonnapark.wnpserver.episode.Episode;
+import com.wonnapark.wnpserver.episode.EpisodeFixtures;
 import com.wonnapark.wnpserver.episode.application.EpisodeFindUseCase;
 import com.wonnapark.wnpserver.episode.dto.response.EpisodeDetailFormResponse;
 import com.wonnapark.wnpserver.episode.dto.response.EpisodeListFormResponse;
-import com.wonnapark.wnpserver.episode.presentation.UserEpisodeController;
 import com.wonnapark.wnpserver.global.auth.AuthFixtures;
 import com.wonnapark.wnpserver.global.auth.Authentication;
 import com.wonnapark.wnpserver.global.auth.AuthorizedArgumentResolver;
@@ -27,10 +27,6 @@ import java.util.List;
 
 import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.document;
 import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.resourceDetails;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createEpisode;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createEpisodes;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createPageable;
-import static com.wonnapark.wnpserver.domain.episode.EpisodeFixtures.createWebtoon;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
@@ -75,8 +71,8 @@ class UserEpisodeControllerTest {
     void findEpisodeDetailForm() throws Exception {
         // given
         Long episodeId = 1L;
-        Webtoon webtoon = createWebtoon();
-        Episode episode = createEpisode(webtoon);
+        Webtoon webtoon = EpisodeFixtures.createWebtoon();
+        Episode episode = EpisodeFixtures.createEpisode(webtoon);
         given(episodeFindUseCase.findEpisodeDetailForm(userInfo.userId(), episodeId)).willReturn(EpisodeDetailFormResponse.from(episode));
         // when // then
         this.mockMvc.perform(get("/api/v1/user/episode/{episodeId}/detail", episodeId)
@@ -101,9 +97,9 @@ class UserEpisodeControllerTest {
     @DisplayName("유저는 웹툰 ID로 에피소드 리스트 정보를 정상적으로 가져올 수 있다.")
     void findEpisodeListForm() throws Exception {
         // given
-        Pageable pageable = createPageable();
-        Webtoon webtoon = createWebtoon();
-        List<Episode> episodes = createEpisodes(webtoon);
+        Pageable pageable = EpisodeFixtures.createPageable();
+        Webtoon webtoon = EpisodeFixtures.createWebtoon();
+        List<Episode> episodes = EpisodeFixtures.createEpisodes(webtoon);
         given(episodeFindUseCase.findEpisodeListForm(userInfo.userId(), webtoon.getId(), pageable))
                 .willReturn(new PageImpl<>(episodes, pageable, episodes.size()).map(EpisodeListFormResponse::from));
         // when // then
