+++ wnp-server/src/test/java/com/wonnapark/wnpserver/webtoon/application/GuestWebtoonServiceTest.java
@@ -1,8 +1,7 @@
-package com.wonnapark.wnpserver.domain.webtoon.application;
+package com.wonnapark.wnpserver.webtoon.application;
 
 import com.wonnapark.wnpserver.webtoon.Webtoon;
-import com.wonnapark.wnpserver.domain.webtoon.WebtoonFixtures;
-import com.wonnapark.wnpserver.webtoon.application.GuestWebtoonService;
+import com.wonnapark.wnpserver.webtoon.WebtoonFixtures;
 import com.wonnapark.wnpserver.webtoon.dto.response.WebtoonDetailResponse;
 import com.wonnapark.wnpserver.webtoon.exception.UnderageAccessDeniedException;
 import com.wonnapark.wnpserver.webtoon.infrastructure.WebtoonRepository;
